API working/requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.1
Faker==4.1.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
